+++ Source/Utility.cpp
@@ -67,12 +67,12 @@ sf::Vector2i PixelToChessBoard(sf::Vector2i posPixel, sf::Vector2i start, sf::Ve
 sf::Vector2i CoordToPixelChessBoard(sf::Vector2i posBoard, sf::Vector2i start, sf::Vector2i offset, bool reverse) {
 	int posX, posY;
 	if (reverse) {
-		posX = start.x + (float)(7 - posBoard.x) * offset.x;
-		posY = start.y + (float)posBoard.y * offset.y;
+		posX = start.x + (7 - posBoard.x) * offset.x;
+		posY = start.y + posBoard.y * offset.y;
 	}
 	else {
-		posX = start.x + (float)posBoard.x * offset.x;
-		posY = start.y + (float)(7 - posBoard.y) * offset.y;
+		posX = start.x + posBoard.x * offset.x;
+		posY = start.y + (7 - posBoard.y) * offset.y;
 	}
 	return sf::Vector2i(posX, posY);
 }
